Переименовал компоненты и импорты
