+++ project/src/components/main/films-list.tsx
