+++ project/src/components/player/preview-player.tsx
